Move runtime related function into runtime.rs
