+++ src/build.rs
@@ -1,81 +1,16 @@
-use std::error::Error;
-use std::{env, fs, path::PathBuf, process};
+use std::{fs, path::PathBuf};
 
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 
-use crate::data::layer_content_metadata::LayerContentMetadata;
-use crate::error::LibCnbError;
-use crate::shared::{write_toml_file, TomlFileError};
 use crate::{
-    data::{buildpack::BuildpackToml, buildpack_plan::BuildpackPlan, launch::Launch},
+    data::{
+        buildpack::BuildpackToml, buildpack_plan::BuildpackPlan, launch::Launch,
+        layer_content_metadata::LayerContentMetadata,
+    },
     platform::Platform,
-    shared::read_toml_file,
+    shared::{read_toml_file, write_toml_file, TomlFileError},
 };
-
-pub fn cnb_runtime_build<
-    E: Error,
-    F: Fn(BuildContext<P, BM>) -> Result<(), LibCnbError<E>>,
-    BM: DeserializeOwned,
-    P: Platform,
->(
-    build_fn: F,
-) -> Result<(), LibCnbError<E>> {
-    let args = parse_build_args_or_exit();
-
-    let layers_dir = args.layers_dir_path;
-
-    let app_dir = env::current_dir().map_err(LibCnbError::CannotDetermineAppDirectory)?;
-
-    let buildpack_dir = env::var("CNB_BUILDPACK_DIR")
-        .map_err(LibCnbError::CannotDetermineBuildpackDirectory)
-        .map(PathBuf::from)?;
-
-    let stack_id: String = env::var("CNB_STACK_ID").map_err(LibCnbError::CannotDetermineStackId)?;
-
-    let platform =
-        P::from_path(&args.platform_dir_path).map_err(LibCnbError::CannotCreatePlatformFromPath)?;
-
-    let buildpack_plan =
-        read_toml_file(&args.buildpack_plan_path).map_err(LibCnbError::CannotReadBuildpackPlan)?;
-
-    let buildpack_descriptor = read_toml_file(buildpack_dir.join("buildpack.toml"))
-        .map_err(LibCnbError::CannotReadBuildpackDescriptor)?;
-
-    let context = BuildContext {
-        layers_dir,
-        app_dir,
-        buildpack_dir,
-        stack_id,
-        platform,
-        buildpack_plan,
-        buildpack_descriptor,
-    };
-
-    build_fn(context)
-}
-
-struct BuildArgs {
-    pub layers_dir_path: PathBuf,
-    pub platform_dir_path: PathBuf,
-    pub buildpack_plan_path: PathBuf,
-}
-
-fn parse_build_args_or_exit() -> BuildArgs {
-    let args: Vec<String> = env::args().collect();
-    match args.as_slice() {
-        [_, layers_dir_path, platform_dir_path, buildpack_plan_path] => BuildArgs {
-            layers_dir_path: PathBuf::from(layers_dir_path),
-            platform_dir_path: PathBuf::from(platform_dir_path),
-            buildpack_plan_path: PathBuf::from(buildpack_plan_path),
-        },
-        _ => {
-            eprintln!("Usage: build <layers> <platform> <plan>");
-            eprintln!("https://github.com/buildpacks/spec/blob/main/buildpack.md#build");
-            process::exit(1);
-        }
-    }
-}
 
 pub struct BuildContext<P: Platform, BM> {
     pub layers_dir: PathBuf,
